Enum scad_generator::CircleType

pub enum CircleType {
    Radius(f32),
    Diameter(f32),
}

Since scad allows creation of circle like objects using either radius or diameter, this enum specifies which format to use

Variants
Radius(f32)Diameter(f32)