Enum scad_generator::ScadElement

pub enum ScadElement {
    Translate(Vector3<f32>),
    Scale(Vector3<f32>),
    Rotate(f32, Vector3<f32>),
    Mirror(Vector3<f32>),
    LinearExtrude(LinExtrudeParams),
    Difference,
    Union,
    Hull,
    Intersection,
    Cube(Vector3<f32>),
    Cylinder(f32, CircleType),
    Sphere(CircleType),
    Cone(f32, CircleType, CircleType),
    Polyhedron(Vec<Vector3<f32>>, Vec<Vec<i32>>),
    Polygon(Vec<Vector2<f32>>, Vec<Vec<i32>>),
    Import(String),
    Square(Vector2<f32>),
}

Different kinds of scad modules and function. These are parameters for ScadObjects.

Most of these have the same name as the openscad counterparts so see their documentation for details

Variants
Translate(Vector3<f32>)Scale(Vector3<f32>)Rotate(f32, Vector3<f32>)Mirror(Vector3<f32>)LinearExtrude(LinExtrudeParams)DifferenceUnionHullIntersectionCube(Vector3<f32>)Cylinder(f32, CircleType)Sphere(CircleType)Cone(f32, CircleType, CircleType)Polyhedron(Vec<Vector3<f32>>, Vec<Vec<i32>>)Polygon(Vec<Vector2<f32>>, Vec<Vec<i32>>)Import(String)Square(Vector2<f32>)
Methods

impl ScadElement

fn get_code(self) -> String

Returns scad code for each of the elements