Struct scad_generator::LinExtrudeParams

pub struct LinExtrudeParams {
    pub height: f32,
    pub center: bool,
    pub convexity: i32,
    pub twist: f32,
    pub slices: i32,
}

Parameters for the linear extrude function.

These are in a struct because there are so many of them and most of them can have a default value.

Fields
height: f32 center: bool convexity: i32 twist: f32 slices: i32