Struct scad_generator::ScadFile

pub struct ScadFile { /* fields omitted */ }

Object that stores scad objects along with global parameters for the objects. Also has methods for writing the data to files.

Methods

impl ScadFile

fn new() -> ScadFile

fn get_code(&self) -> String

Returns the code for the global parameters as well as all the children in the file

fn add_object(&mut self, object: ScadObject)

fn set_detail(&mut self, detail: i32)

Sets the $fn variable for the whole file. This varibale defines the detail amount for cylindrical objects

fn write_to_file(&self, path: String) -> bool

Writes the resulting code to a file

Arguments

path: The path to the file where we want to write relative to the current working directory.

Returns

The function will return false and print a message to the console if writing fails.